requirements.txt
click==8.0.3
Flask==2.0.2
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
protobuf==3.19.1
watchdog==2.1.6
Werkzeug==2.0.2
